src/backend/ai-engine/requirements.txt
# FILE: src/ai-engine/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
librosa
python-multipart
numpy
soundfile
gTTS
